Prevent a crash loop caused by upstream changes
csi-lib-utils.Connect now times out and returns an error after 30 seconds by
default. If the CSI plugin we are connecting to is unavailable during
livenessprobe initialization, livenessprobe will crash and potentially enter a
crash loop. If the backoff on both the CSI plugin and livenessprobe becomes
large, the situation is unrecoverable.

Configure the timeout of the csi-lib-utils.Connect function and remove
acquireConnection since the upstream changes have made it obselete.

Signed-off-by: Eric Weber <eric.weber@suse.com>