Don't exit the probe on connection issues
Do not exit the liveness probe process when the liveness probe cannot
connect to the CSI driver. The driver could be crashlooping, and we should
not crashloop the liveness probe process too. The process should only fail
all probes to /healthz endpoint. Since the HTTP server is not running when
connecting to the driver for the first time, "connection refused" must
be a good enough failure.